+++ ingsoc-js/index.js
@@ -1,9 +1,9 @@
 #! /usr/bin/env node
 import { program } from 'commander';
-import build from './lib/build.js';
-import serve from './lib/serve.js';
+import build from './lib/cli//build.js';
+import serve from './lib/cli/serve.js';
 
-program.version('0.0.1');
+program.version('1.0.0');
 
 program.command('build')
     .description('Build')
