+++ ingsoc-js/lib/cli/build-logic/templates.js
