+++ ingsoc-js/lib/newspeak/ingsoc.js
