+++ ingsoc-js/lib/newspeak/thinkpol-callbacks.js
