+++ ingsoc-js/lib/newspeak/thinkpol.js
