+++ ingsoc-js/lib/newspeak/thoughtcrime.js
